% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctRatio.R
\name{correctRatio}
\alias{correctRatio}
\title{correctRatio}
\usage{
correctRatio(x, y, top = 3)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
correct or wrong.
}
\description{
Calculate top n correction.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
