% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decideCIE.R
\name{decideCIE}
\alias{decideCIE}
\title{decideCIE}
\usage{
decideCIE(
  id1,
  id2,
  top1 = 1,
  top2 = 3,
  id.type = c("kegg", "hmdb", "cas"),
  library = NULL
)
}
\arguments{
\item{id1}{Standard IDs.}

\item{id2}{Identifications from new software.}

\item{top1}{Top n for ID1.}

\item{top2}{Top n for ID2.}

\item{id.type}{the ID type, KEGG, HMDB or CAS.}
}
\value{
Correct, isomer and wrong.
}
\description{
Decide it is correct, isomer or error identification.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
