% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupSet.R
\name{dupSet}
\alias{dupSet}
\title{dupSet}
\usage{
dupSet(data)
}
\arguments{
\item{data}{A maxtrix or data.frame.}
}
\value{
like duplicated.
}
\description{
dupSet.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
