% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errordDisplay.R
\name{errorDisplay}
\alias{errorDisplay}
\title{errordDisplay}
\usage{
errorDisplay(expr, error.info = "error")
}
\arguments{
\item{expr}{An R expression.}

\item{error.info}{The error information you want to display.}
}
\value{
Detailed expression information.
}
\description{
Displary the error.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
