% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWorklist.R
\name{getWorklist}
\alias{getWorklist}
\title{getWorklist}
\usage{
getWorklist(instrument = c("Thermo", "Agilent", "AB"),
  polarity = c("both", "positive", "negative"), each.mode.number = 30,
  name = "worklist", randommethod = c("no", "position", "injection"),
  samplenumber = NULL, replication = 1, QCstep = 6,
  conditionQCnumber = 5, validationQCstep = 40, testmixstep = 40,
  injectionfrom = 1, user = "other",
  dir = "D:\\\\Liang\\\\SMARD-URINE-20190701\\\\HILIC\\\\")
}
\arguments{
\item{instrument}{Which instrument you use?
"Agilent", "AB" or "Thermo", default is "Thermo".}

\item{name}{The name of worklist.}

\item{randommethod}{Which random method you want to use?
"no", "position" or "injection". Default is "no".}

\item{samplenumber}{Sample number.}

\item{replication}{Replication times.}

\item{QCstep}{QC step.}

\item{conditionQCnumber}{Condition QC number.}

\item{validationQCstep}{Validation QC step.}

\item{testmixstep}{Test mixture step.}

\item{injectionfrom}{Injection order from which? Default is 1.}

\item{user}{Default is "other".}

\item{dir}{Directory.}
}
\value{
New worklist.
}
\description{
Generate Worklist for MS data acquisition.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
