% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installBioc.R
\name{installBioc}
\alias{installBioc}
\title{installBioc}
\usage{
installBioc(package = "Rdisop", web = c("https", "http"), ...)
}
\arguments{
\item{package}{The package name from Bioconductor.}

\item{...}{Other parameters. See ?biocLite}
}
\description{
Install packages from Bioconductor.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}

Xiaotao Shen
}
