% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxtMTmatch.R
\name{keepOne}
\alias{keepOne}
\title{keepOne}
\usage{
keepOne(result, according.to = c("mz.error", "rt.error"))
}
\arguments{
\item{result}{result from sxtMTmatch function.}

\item{according.to}{According to mz error or rt error?}
}
\value{
Return a result without multiple vs. one.
}
\description{
Remove multiple vs. one in result from sxtMTmatch function.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
