% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Plot.R
\name{ms2Plot}
\alias{ms2Plot}
\title{ms2Plot}
\usage{
ms2Plot(
  spectrum1,
  spectrum2,
  col1 = "lightseagreen",
  col2 = "tomato",
  col3 = "grey",
  xlim = NULL,
  lwd = 2,
  cex.lab = 1.8,
  cex.axis = 1.5,
  xlab = "m/z",
  ylab = "Relative intensity",
  real.int.cutoff = 0,
  ppm.ms2match = 30,
  bty = "o",
  xaxs = "r",
  yaxs = "r",
  ...
)
}
\arguments{
\item{spectrum1}{Spectrum 1.}

\item{spectrum2}{Spectrum 2.}

\item{col1}{Color 1.}

\item{col2}{Color 2.}

\item{col3}{Color 3.}

\item{xlim}{xlim.}

\item{lwd}{lwd.}

\item{cex.lab}{cex.lab.}

\item{cex.axis}{cex.axis.}

\item{xlab}{xlab.}

\item{ylab}{ylab.}

\item{real.int.cutoff}{real.int.cutoff.}

\item{ppm.ms2match}{ppm.ms2match.}

\item{bty}{bty}

\item{xaxs}{xaxs}

\item{yaxs}{yaxs}

\item{...}{Other parameters.}
}
\value{
Return a MS2 spectrum.
}
\description{
Plot ms2 spectrum.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
