% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Plot.R
\name{ms2Plot2}
\alias{ms2Plot2}
\title{ms2Plot2}
\usage{
ms2Plot2(spectrum1, spectrum2, range.mz, ppm.tol = 30,
  mz.ppm.thr = 400, xlab = "Mass to charge ratio (m/z)",
  ylab = "Relative intensity", col1 = "red", col2 = "black",
  title.size = 15, lab.size = 15, axis.text.size = 15,
  legend.title.size = 15, legend.text.size = 15)
}
\arguments{
\item{spectrum1}{Spectrum 1.}

\item{spectrum2}{Spectrum 2.}

\item{range.mz}{range.mz}

\item{ppm.tol}{ppm.tol}

\item{mz.ppm.thr}{mz.ppm.thr}

\item{xlab}{xlab.}

\item{ylab}{ylab.}

\item{col1}{Color 1.}

\item{col2}{Color 2.}

\item{title.size}{title.size}

\item{axis.text.size}{axis.text.size.}

\item{legend.title.size}{legend.title.size}

\item{legend.text.size}{legend.text.size}

\item{real.int.cutoff}{real.int.cutoff.}
}
\value{
Return a MS2 spectrum.
}
\description{
Plot ms2 spectrum.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
