% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeIsotope.R
\name{removeIsotope}
\alias{removeIsotope}
\title{removeIsotope}
\usage{
removeIsotope(data, data.from = c("metdna", "metanalyzer", "xcms"))
}
\arguments{
\item{data}{A maxtrix or data.frame.}

\item{data.from}{data from where?}
}
\value{
A data without isotopes.
}
\description{
removeIsotope
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
