% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxtCorTest.R
\name{sxtCorTest}
\alias{sxtCorTest}
\title{sxtCorTest}
\usage{
sxtCorTest(x, alternative = c("two.sided", "less", "greater"),
  method = c("pearson", "kendall", "spearman"), exact = NULL,
  conf.level = 0.95, continuity = FALSE)
}
\arguments{
\item{data}{A maxtrix or data.frame.}
}
\value{
like cor.
}
\description{
sxtCorTest
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
