% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxtRSD.R
\name{sxtRSD}
\alias{sxtRSD}
\title{sxtRSD}
\usage{
sxtRSD(x, rm.na = FALSE)
}
\arguments{
\item{x}{A vector of number.}

\item{rm.na}{Remove NA or not.}
}
\description{
Calculate RSD.
}
\author{
Xiaotao Shen
}
