% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxtScale.R
\name{sxtScale}
\alias{sxtScale}
\title{sxtScale}
\usage{
sxtScale(data, method = c("no", "pareto", "auto"), center = TRUE)
}
\arguments{
\item{data}{A matarix or data frame, rows represents variables, and columns represents observations.}

\item{method}{Scale method. no, pareto or auto.}

\item{center}{Center data or not, default is TRUE.}
}
\value{
A scaled data.
}
\description{
Scale data (matirx or data frame)
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
