% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxtSection.R
\name{sxtSection}
\alias{sxtSection}
\title{sxtSection}
\usage{
sxtSection(vector, n = 1)
}
\arguments{
\item{vector}{A vector.}

\item{n}{Length.}
}
\value{
Sected vector.
}
\description{
Sect vector.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
