% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxtVolcanoPlot.R
\name{sxtVolcanoPlot}
\alias{sxtVolcanoPlot}
\title{sxtVolcanoPlot}
\usage{
sxtVolcanoPlot(
  p.value,
  fc,
  p.cutoff = 0.05,
  fc.cutoff = 2,
  linetype = c("3", "2", "1"),
  xlab = "log2(Fold change)",
  ylab = "-log10(P-value, FDR)",
  title = "Volcano plot",
  legend.title = "Marker",
  legend.label = c("Yes", "No"),
  point.color = c("#F8766D", "#619CFF"),
  alpha = 0.9,
  lab.size = 12,
  axis.text.size = 12,
  title.size = 15,
  legend.title.size = 12,
  legend.text.size = 10,
  ...
)
}
\arguments{
\item{p.value}{A numeric vector for p values.}

\item{fc}{A numeric vector for fold change values.}

\item{p.cutoff}{The cutoff of p values.}

\item{fc.cutoff}{The cutoff of fold change values, only set the larger side.}

\item{linetype}{The line type for cutoff of p values and fold change values.}

\item{xlab}{Title of x axis.}

\item{ylab}{Title of y axis.}

\item{title}{Title of plot.}

\item{legend.title}{Legend title.}

\item{legend.label}{Legend label.}

\item{point.color}{Colors of points.}

\item{alpha}{Alpha of points.}

\item{lab.size}{Font size of axis titles.}

\item{axis.text.size}{Font size of axis text.}

\item{title.size}{Font size of title.}

\item{legend.title.size}{Font size of legend title.}

\item{legend.text.size}{Font size of axis text.}
}
\value{
A ggplot object.
}
\description{
Volcano plot
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
