% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_theme.R
\name{theme_sxtlight}
\alias{theme_sxtlight}
\title{theme_sxtlight}
\usage{
theme_sxtlight(...)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
correct or wrong.
}
\description{
Calculate top n correction.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
